requirements.txt
Django==4.2.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==11.1.0
schedule==1.2.0